agno_tutorial/0_requirements.txt
openai 
agno
duckduckgo-search
lancedb
tantivy
pypdf
yfinance
python-dotenv
sqlalchemy
fastapi[standard]